Move TransactionInfo to common folder
