+++ packages/ui/src/memberships/modals/InviteMemberModal/InviteMemberSignModal.tsx
@@ -5,11 +5,11 @@ import React, { useEffect, useState } from 'react'
 import { SelectedAccount } from '../../../accounts/components/SelectAccount'
 import { useAccounts } from '../../../accounts/hooks/useAccounts'
 import { useBalance } from '../../../accounts/hooks/useBalance'
-import { TransactionInfo } from '../../../accounts/modals/TransferModal/TransactionInfo'
 import { accountOrNamed } from '../../../accounts/model/accountOrNamed'
 import { ButtonPrimary } from '../../../common/components/buttons'
 import { InputComponent } from '../../../common/components/forms'
 import { ModalBody, ModalFooter, Row, TransactionInfoContainer } from '../../../common/components/Modal'
+import { TransactionInfo } from '../../../common/components/TransactionInfo'
 import { TransactionModal } from '../../../common/components/TransactionModal'
 import { TextMedium, TokenValue } from '../../../common/components/typography'
 import { useSignAndSendTransaction } from '../../../common/hooks/useSignAndSendTransaction'
