+++ packages/ui/src/memberships/modals/UpdateMembershipModal/UpdateMembershipSignModal.tsx
@@ -3,11 +3,11 @@ import { SubmittableExtrinsic } from '@polkadot/api/types'
 import React, { useMemo } from 'react'
 
 import { SelectedAccount } from '../../../accounts/components/SelectAccount'
-import { TransactionInfo } from '../../../accounts/modals/TransferModal/TransactionInfo'
 import { Account } from '../../../accounts/types'
 import { ButtonPrimary } from '../../../common/components/buttons'
 import { Label } from '../../../common/components/forms'
 import { ModalBody, ModalFooter, Row, TransactionInfoContainer } from '../../../common/components/Modal'
+import { TransactionInfo } from '../../../common/components/TransactionInfo'
 import { TransactionModal } from '../../../common/components/TransactionModal'
 import { TextMedium, TokenValue } from '../../../common/components/typography'
 import { useApi } from '../../../common/hooks/useApi'
